fix central tx processor
