+++ src/entrypoints/rfp/mod.rs
@@ -1,7 +1,6 @@
 use self::rfp_types::*;
-use crate::db::db_types::{RfpSnapshotRecord, RfpWithLatestSnapshotView};
+use crate::db::db_types::RfpWithLatestSnapshotView;
 use crate::db::DB;
-use crate::nearblocks_client::types::Transaction;
 use crate::rpc_service::RpcService;
 use crate::types::{Contract, PaginatedResponse};
 use crate::{nearblocks_client, timestamp_to_date_string};
@@ -19,20 +18,6 @@ pub mod rfp_types;
 // fn search(input: String) -> Option<Json<PaginatedResponse<RfpWithLatestSnapshotView>>> {
 //     None
 // }
-
-fn get_rfp_id(transaction: &Transaction) -> Result<i32, &'static str> {
-    let action = transaction
-        .actions
-        .first()
-        .ok_or("No actions found in transaction")?;
-
-    let args: PartialEditRFPArgs = serde_json::from_str(&action.args).map_err(|e| {
-        eprintln!("Failed to parse JSON: {:?}", e);
-        "Failed to parse proposal arguments"
-    })?;
-
-    Ok(args.id)
-}
 
 #[utoipa::path(get, path = "/rfps?<order>&<limit>&<offset>&<filters>", params(
   ("order"= &str, Path, description ="order"),
@@ -108,7 +93,7 @@ async fn get_rfps(
             nearblocks_client::ApiResponse { txns: vec![] }
         }
     };
-    let _ = process_transactions(&nearblocks_unwrapped.txns, db).await;
+    let _ = nearblocks_client::transactions::process(&nearblocks_unwrapped.txns, db).await;
 
     let (rfps, total) = match db
         .get_rfps_with_latest_snapshot(limit, order, offset, filters)
@@ -127,115 +112,6 @@ async fn get_rfps(
         limit.try_into().unwrap(),
         total.try_into().unwrap(),
     )))
-}
-
-async fn process_transactions(transactions: &[Transaction], db: &State<DB>) -> Result<(), Status> {
-    for transaction in transactions.iter() {
-        if let Some(action) = transaction.actions.first() {
-            let result = match action.method.as_str() {
-                "set_rfp_block_height_callback" => {
-                    handle_set_rfp_block_height_callback(transaction.to_owned(), db).await
-                }
-                "edit_proposal_versioned_timeline" => {
-                    handle_edit_rfp(transaction.to_owned(), db).await
-                }
-                "edit_proposal_timeline" => handle_edit_rfp(transaction.to_owned(), db).await,
-                "edit_proposal" => handle_edit_rfp(transaction.to_owned(), db).await,
-                "edit_proposal_linked_rfp" => handle_edit_rfp(transaction.to_owned(), db).await,
-                _ => {
-                    println!("Unhandled method: {}", action.method);
-                    continue;
-                }
-            };
-            result?;
-        }
-    }
-
-    Ok(())
-}
-
-async fn handle_set_rfp_block_height_callback(
-    transaction: Transaction,
-    db: &State<DB>,
-) -> Result<(), Status> {
-    let action = transaction.clone().actions.first().unwrap().clone();
-    let json_args = action.args.clone();
-
-    // println!("json_args: {:?}", json_args.clone());
-    let args: SetBlockHeightCallbackArgs = serde_json::from_str(&json_args).unwrap();
-
-    println!("Adding to the database... {}", args.clone().rfp.id);
-    let mut tx = db.begin().await.map_err(|_e| Status::InternalServerError)?;
-    DB::upsert_proposal(
-        &mut tx,
-        args.clone().rfp.id,
-        args.clone().rfp.author_id.to_string(),
-    )
-    .await
-    .unwrap();
-
-    let rpc_service = RpcService::default();
-    let id = args.clone().rfp.id.try_into().unwrap();
-
-    let versioned_rfp_fallback: VersionedRFP = args.clone().rfp.into();
-    let versioned_rfp = match rpc_service.get_rfp(id).await {
-        Ok(rfp) => rfp.data,
-        Err(e) => {
-            eprintln!(
-                "Failed to get RFP from RPC, using first snapshot as fallback {:?}",
-                e
-            );
-            versioned_rfp_fallback
-        }
-    };
-
-    let snapshot = RfpSnapshotRecord::from_contract_rfp(
-        versioned_rfp.into(),
-        transaction.block_timestamp,
-        transaction.block.block_height,
-    );
-
-    DB::insert_rfp_snapshot(&mut tx, &snapshot).await.unwrap();
-
-    tx.commit()
-        .await
-        .map_err(|_e| Status::InternalServerError)?;
-
-    Ok(())
-}
-
-async fn handle_edit_rfp(transaction: Transaction, db: &State<DB>) -> Result<(), Status> {
-    let rpc_service = RpcService::default();
-    let id = get_rfp_id(&transaction).map_err(|e| {
-        eprintln!("Failed to get RFP ID: {}", e);
-        Status::InternalServerError
-    })?;
-
-    let versioned_rfp = match rpc_service.get_rfp(id).await {
-        Ok(rfp) => rfp.data,
-        Err(e) => {
-            eprintln!("Failed to get rfp from RPC: {:?}", e);
-            return Err(Status::InternalServerError);
-        }
-    };
-
-    let mut tx = db.begin().await.map_err(|_e| Status::InternalServerError)?;
-
-    let snapshot = RfpSnapshotRecord::from_contract_rfp(
-        versioned_rfp.into(),
-        transaction.block_timestamp,
-        transaction.block.block_height,
-    );
-
-    DB::insert_rfp_snapshot(&mut tx, &snapshot)
-        .await
-        .map_err(|_e| Status::InternalServerError)?;
-
-    tx.commit()
-        .await
-        .map_err(|_e| Status::InternalServerError)?;
-
-    Ok(())
 }
 
 #[utoipa::path(get, path = "/rfps/{rfp_id}")]
